Refactor ClangUtils::build_info_details to take a cindex module
Instead of a layer of redirection when lib_complete.py builds lists of
types from cindex, just keep a reference to the loaded cindex module
and pass that around. There should be no visible changes from this
commit, but this will make future work on macros easier because there
will more direct access to cindex's types and methods.